fix randint yet again

+++ src/psgd_sampled.c
@@ -140,12 +140,12 @@ inline size_t randint(size_t nmax, unsigned int *seed)
 		do
 		{
 			for (size_t d = 0; d < ndraws; d++) {
-				*((int*)(ptr_drawn + d * sizeof(int)))  = rand_r(seed);
-				*((int*)(ptr_drawn + d * sizeof(int))) += rand_r(seed);
+				*((unsigned int*)(ptr_drawn + d * sizeof(int)))  = (unsigned int) rand_r(seed);
+				*((unsigned int*)(ptr_drawn + d * sizeof(int))) += (unsigned int) rand_r(seed);
 			}
 			if (n_remainder) {
-				single_int  = rand_r(seed);
-				single_int += rand_r(seed);
+				single_int  = (unsigned int) rand_r(seed);
+				single_int += (unsigned int) rand_r(seed);
 				memcpy(ptr_drawn + ndraws * sizeof(int), &single_int, n_remainder);
 			}
 		} while (n > lim);
